feat: onBrushEnd as part of a heatmap config
