+++ src/chart_types/heatmap/state/selectors/is_brush_available.ts
@@ -20,21 +20,17 @@
 import createCachedSelector from 're-reselect';
 
 import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
-import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { getHeatmapConfigSelector } from './get_heatmap_config';
 
 /**
  * The brush is available only if a onBrushEnd listener is configured
  * @internal
  */
-export const isBrushAvailableSelector = createCachedSelector(
-  [getSettingsSpecSelector, getHeatmapConfigSelector],
-  (settingsSpec, config): boolean => {
-    return Boolean(settingsSpec.onBrushEnd) && config.brushTool.visible;
-  },
-)(getChartIdSelector);
+export const isBrushAvailableSelector = createCachedSelector([getHeatmapConfigSelector], (config): boolean => {
+  return Boolean(config.onBrushEnd) && config.brushTool.visible;
+})(getChartIdSelector);
 
 /** @internal */
-export const isBrushEndProvided = createCachedSelector([getSettingsSpecSelector], (settingsSpec): boolean => {
-  return Boolean(settingsSpec.onBrushEnd);
+export const isBrushEndProvided = createCachedSelector([getHeatmapConfigSelector], (config): boolean => {
+  return Boolean(config.onBrushEnd);
 })(getChartIdSelector);
