+++ src/chart_types/xy_chart/state/chart_state.interactions.test.ts
@@ -22,7 +22,7 @@ import { Store } from 'redux';
 import { ChartTypes } from '../..';
 import { MockStore } from '../../../mocks/store';
 import { ScaleType } from '../../../scales/constants';
-import { HeatmapBrushEvent, SettingsSpec, XScaleType, XYBrushArea } from '../../../specs';
+import { SettingsSpec, XScaleType, XYBrushArea } from '../../../specs';
 import { SpecTypes, DEFAULT_SETTINGS_SPEC, TooltipType, BrushAxis } from '../../../specs/constants';
 import { onExternalPointerEvent } from '../../../state/actions/events';
 import { onPointerMove, onMouseDown, onMouseUp } from '../../../state/actions/mouse';
@@ -796,7 +796,7 @@ function mouseOverTestSuite(scaleType: XScaleType) {
   });
   describe('brush', () => {
     test('can respond to a brush end event', () => {
-      const brushEndListener = jest.fn<void, [XYBrushArea | HeatmapBrushEvent]>((): void => undefined);
+      const brushEndListener = jest.fn<void, [XYBrushArea]>((): void => undefined);
       const onBrushCaller = createOnBrushEndCaller();
       store.subscribe(() => {
         onBrushCaller(store.getState());
@@ -881,7 +881,7 @@ function mouseOverTestSuite(scaleType: XScaleType) {
       }
     });
     test('can respond to a brush end event on rotated chart', () => {
-      const brushEndListener = jest.fn<void, [XYBrushArea | HeatmapBrushEvent]>((): void => undefined);
+      const brushEndListener = jest.fn<void, [XYBrushArea]>((): void => undefined);
       const onBrushCaller = createOnBrushEndCaller();
       store.subscribe(() => {
         onBrushCaller(store.getState());
@@ -953,7 +953,7 @@ function mouseOverTestSuite(scaleType: XScaleType) {
       }
     });
     test('can respond to a Y brush', () => {
-      const brushEndListener = jest.fn<void, [XYBrushArea | HeatmapBrushEvent]>((): void => undefined);
+      const brushEndListener = jest.fn<void, [XYBrushArea]>((): void => undefined);
       const onBrushCaller = createOnBrushEndCaller();
       store.subscribe(() => {
         onBrushCaller(store.getState());
@@ -1029,7 +1029,7 @@ function mouseOverTestSuite(scaleType: XScaleType) {
       }
     });
     test('can respond to rectangular brush', () => {
-      const brushEndListener = jest.fn<void, [XYBrushArea | HeatmapBrushEvent]>((): void => undefined);
+      const brushEndListener = jest.fn<void, [XYBrushArea]>((): void => undefined);
       const onBrushCaller = createOnBrushEndCaller();
       store.subscribe(() => {
         onBrushCaller(store.getState());
