+++ src/utils/events.ts
@@ -18,7 +18,7 @@
  */
 
 import { Scale } from '../scales';
-import { BrushEndListener, isPointerOverEvent, PointerEvent, PointerOverEvent } from '../specs';
+import { BrushEndListener, isPointerOverEvent, PointerEvent, PointerOverEvent, HeatmapSpec } from '../specs';
 import { DragState } from '../state/chart_state';
 
 /** @internal */
@@ -31,7 +31,7 @@ export function isValidPointerOverEvent(
 
 /** @internal */
 export interface DragCheckProps {
-  onBrushEnd: BrushEndListener | undefined;
+  onBrushEnd: BrushEndListener | HeatmapSpec['config']['onBrushEnd'] | undefined;
   lastDrag: DragState | null;
 }
